+++ RHILib/shaders/point.vert
@@ -18,7 +18,7 @@ layout(std140, binding = 1) uniform MeshBlock {
     mat4 mvp;
     mat4 mv;
     vec4 col;
-    float useClipping;
+    int useClipping;
 } mesh;
 
 void main()
